Extract helper for handshake message transitions
The extracted class `Handshake_Transitions` will be used independently from the
TLS 1.2 `Handshake_State` in the upcoming TLS 1.3 code base.
